Moved method to chef model instead of dish
