+++ app/views/ingredients/index.html.erb
@@ -1,5 +1,3 @@
 <h1>List of All Ingredients</h1>
 
-<% @chef.dishes.each do |dish| %>
-  <li><%= dish.list_of_ingredients%></li>
-<% end %>
+<li><%= @chef.list_of_ingredients %></li>
